Split users sessions and settings into new modules
- Split duties
