+++ lib/bank/users.ex
@@ -12,28 +12,6 @@ defmodule Bank.Users do
   alias Bank.Users.UserToken
 
   @doc """
-  Emulates that the email will change without actually changing
-  it in the database.
-
-  ## Examples
-
-      iex> apply_user_email(user, "valid password", %{email: ...})
-      {:ok, %User{}}
-
-      iex> apply_user_email(user, "invalid password", %{email: ...})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  @spec apply_user_email(Ecto.Schema.t(), binary(), map()) ::
-          {:ok, Ecto.Schema.t()} | map() | {:error, Ecto.Changeset.t()}
-  def apply_user_email(user, password, attrs) do
-    user
-    |> User.email_changeset(attrs)
-    |> User.validate_current_password(password)
-    |> Ecto.Changeset.apply_action(:update)
-  end
-
-  @doc """
   Returns an `%Ecto.Changeset{}` for tracking user changes.
 
   ## Examples
@@ -43,20 +21,6 @@ defmodule Bank.Users do
   """
   @spec change_user(Ecto.Schema.t(), map()) :: Ecto.Changeset.t()
   def change_user(%User{} = user, attrs \\ %{}), do: User.changeset(user, attrs)
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for changing the user email.
-
-  ## Examples
-
-      iex> change_user_email(user)
-      %Ecto.Changeset{data: %User{}}
-
-  """
-  @spec change_user_email(Ecto.Schema.t(), map()) :: Ecto.Changeset.t()
-  def change_user_email(user, attrs \\ %{}) do
-    User.email_changeset(user, attrs, validate_email: false)
-  end
 
   @doc """
   Returns an `%Ecto.Changeset{}` for changing the user password.
@@ -87,24 +51,6 @@ defmodule Bank.Users do
   end
 
   @doc """
-  Confirms a user by the given token.
-
-  If the token matches, the user account is marked as confirmed
-  and the token is deleted.
-  """
-  @spec confirm_user(binary()) :: {:ok, Ecto.Schema.t()} | :error
-  def confirm_user(token) do
-    with {:ok, query} <- UserToken.verify_email_token_query(token, "confirm"),
-         %User{} = user <- Repo.one(query),
-         # credo:disable-for-next-line
-         {:ok, %{user: user}} <- user |> confirm_user_multi() |> Repo.transaction() do
-      {:ok, user}
-    else
-      _ -> :error
-    end
-  end
-
-  @doc """
   Deletes a user.
 
   ## Examples
@@ -118,45 +64,6 @@ defmodule Bank.Users do
   """
   @spec delete_user(Ecto.Schema.t()) :: EctoUtils.write()
   def delete_user(%User{} = user), do: Repo.delete(user)
-
-  @doc """
-  Deletes the signed token with the given context.
-  """
-  @spec delete_user_session_token(binary()) :: :ok
-  def delete_user_session_token(token) do
-    token
-    |> UserToken.by_token_and_context_query("session")
-    |> Repo.delete_all()
-
-    :ok
-  end
-
-  ## Confirmation
-
-  @doc ~S"""
-  Delivers the confirmation email instructions to the given user.
-
-  ## Examples
-
-      iex> deliver_user_confirmation_instructions(user, &url(~p"/users/confirm/#{&1}"))
-      {:ok, %{to: ..., body: ...}}
-
-      iex> deliver_user_confirmation_instructions(confirmed_user, &url(~p"/users/confirm/#{&1}"))
-      {:error, :already_confirmed}
-
-  """
-  @spec deliver_user_confirmation_instructions(Ecto.Schema.t(), fun()) ::
-          {:error, :already_confirmed} | UserNotifier.t()
-  def deliver_user_confirmation_instructions(%User{} = user, confirmation_url_fun)
-      when is_function(confirmation_url_fun, 1) do
-    if user.confirmed_at do
-      {:error, :already_confirmed}
-    else
-      {encoded_token, user_token} = UserToken.build_email_token(user, "confirm")
-      Repo.insert!(user_token)
-      UserNotifier.deliver_confirmation_instructions(user, confirmation_url_fun.(encoded_token))
-    end
-  end
 
   @doc ~S"""
   Delivers the reset password email to the given user.
@@ -173,35 +80,6 @@ defmodule Bank.Users do
     {encoded_token, user_token} = UserToken.build_email_token(user, "reset_password")
     Repo.insert!(user_token)
     UserNotifier.deliver_reset_password_instructions(user, reset_password_url_fun.(encoded_token))
-  end
-
-  @doc ~S"""
-  Delivers the update email instructions to the given user.
-
-  ## Examples
-
-      iex> deliver_user_update_email_instructions(user, current_email, &url(~p"/users/settings/confirm_email/#{&1}"))
-      {:ok, %{to: ..., body: ...}}
-
-  """
-  @spec deliver_user_update_email_instructions(Ecto.Schema.t(), binary(), fun()) ::
-          UserNotifier.t()
-  def deliver_user_update_email_instructions(%User{} = user, current_email, update_email_url_fun)
-      when is_function(update_email_url_fun, 1) do
-    {encoded_token, user_token} = UserToken.build_email_token(user, "change:#{current_email}")
-
-    Repo.insert!(user_token)
-    UserNotifier.deliver_update_email_instructions(user, update_email_url_fun.(encoded_token))
-  end
-
-  @doc """
-  Generates a session token.
-  """
-  @spec generate_user_session_token(Ecto.Schema.t()) :: binary()
-  def generate_user_session_token(user) do
-    {token, user_token} = UserToken.build_session_token(user)
-    Repo.insert!(user_token)
-    token
   end
 
   @doc """
@@ -280,15 +158,6 @@ defmodule Bank.Users do
   end
 
   @doc """
-  Gets the user with the given signed token.
-  """
-  @spec get_user_by_session_token(binary()) :: Ecto.Schema.t() | nil
-  def get_user_by_session_token(token) do
-    {:ok, query} = UserToken.verify_session_token_query(token)
-    Repo.one(query)
-  end
-
-  @doc """
   Registers a user.
 
   ## Examples
@@ -335,59 +204,6 @@ defmodule Bank.Users do
   end
 
   @doc """
-  Updates the user email using the given token.
-
-  If the token matches, the user email is updated and the token is deleted.
-  The confirmed_at date is also updated to the current time.
-  """
-
-  @spec update_user_email(Ecto.Schema.t(), binary()) :: :ok | :error
-  def update_user_email(user, token) do
-    context = "change:#{user.email}"
-
-    with {:ok, query} <- UserToken.verify_change_email_token_query(token, context),
-         %UserToken{sent_to: email} <- Repo.one(query),
-         # credo:disable-for-next-line
-         {:ok, _} <- user |> user_email_multi(email, context) |> Repo.transaction() do
-      :ok
-    else
-      _ -> :error
-    end
-  end
-
-  @doc """
-  Updates the user password.
-
-  ## Examples
-
-      iex> update_user_password(user, "valid password", %{password: ...})
-      {:ok, %User{}}
-
-      iex> update_user_password(user, "invalid password", %{password: ...})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  @spec update_user_password(Ecto.Schema.t(), binary(), map()) ::
-          {:ok, Ecto.Schema.t()} | {:error, Ecto.Changeset.t()}
-  def update_user_password(user, password, attrs) do
-    changeset =
-      user
-      |> User.password_changeset(attrs)
-      |> User.validate_current_password(password)
-
-    ecto_multi =
-      Ecto.Multi.new()
-      |> Ecto.Multi.update(:user, changeset)
-      |> Ecto.Multi.delete_all(:tokens, UserToken.by_user_and_contexts_query(user, :all))
-      |> Repo.transaction()
-
-    case ecto_multi do
-      {:ok, %{user: user}} -> {:ok, user}
-      {:error, :user, changeset, _} -> {:error, changeset}
-    end
-  end
-
-  @doc """
   Returns the list of users.
 
   ## Examples
@@ -416,28 +232,5 @@ defmodule Bank.Users do
     user
     |> User.changeset(attrs)
     |> Repo.update()
-  end
-
-  #
-  # Private functions
-  #
-
-  @spec user_email_multi(Ecto.Schema.t(), binary(), term()) :: Ecto.Multi.t()
-  defp user_email_multi(user, email, context) do
-    changeset =
-      user
-      |> User.email_changeset(%{email: email})
-      |> User.confirm_changeset()
-
-    Ecto.Multi.new()
-    |> Ecto.Multi.update(:user, changeset)
-    |> Ecto.Multi.delete_all(:tokens, UserToken.by_user_and_contexts_query(user, [context]))
-  end
-
-  @spec confirm_user_multi(Ecto.Schema.t()) :: Ecto.Multi.t()
-  defp confirm_user_multi(user) do
-    Ecto.Multi.new()
-    |> Ecto.Multi.update(:user, User.confirm_changeset(user))
-    |> Ecto.Multi.delete_all(:tokens, UserToken.by_user_and_contexts_query(user, ["confirm"]))
   end
 end
