+++ test/bank_web/user_auth_test.exs
@@ -1,8 +1,8 @@
 defmodule BankWeb.UserAuthTest do
   use BankWeb.ConnCase, async: true
 
-  alias Bank.Users
   alias Bank.UsersFixtures
+  alias Bank.UsersSessions
   alias BankWeb.UserAuth
   alias Phoenix.LiveView
 
@@ -23,7 +23,7 @@ defmodule BankWeb.UserAuthTest do
       assert token = get_session(conn, :user_token)
       assert get_session(conn, :live_socket_id) == "users_sessions:#{Base.url_encode64(token)}"
       assert redirected_to(conn) == ~p"/"
-      assert Users.get_user_by_session_token(token)
+      assert UsersSessions.get_user_by_session_token(token)
     end
 
     test "clears everything previously stored in the session", %{conn: conn, user: user} do
@@ -60,7 +60,7 @@ defmodule BankWeb.UserAuthTest do
 
   describe "logout_user/1" do
     test "erases session and cookies", %{conn: conn, user: user} do
-      user_token = Users.generate_user_session_token(user)
+      user_token = UsersSessions.generate_user_session_token(user)
 
       conn =
         conn
@@ -73,7 +73,7 @@ defmodule BankWeb.UserAuthTest do
       refute conn.cookies[@remember_me_cookie]
       assert %{max_age: 0} = conn.resp_cookies[@remember_me_cookie]
       assert redirected_to(conn) == ~p"/"
-      refute Users.get_user_by_session_token(user_token)
+      refute UsersSessions.get_user_by_session_token(user_token)
     end
 
     test "broadcasts to the given live_socket_id", %{conn: conn} do
@@ -101,7 +101,7 @@ defmodule BankWeb.UserAuthTest do
 
   describe "fetch_current_user/2" do
     test "authenticates user from session", %{conn: conn, user: user} do
-      user_token = Users.generate_user_session_token(user)
+      user_token = UsersSessions.generate_user_session_token(user)
 
       conn =
         conn
@@ -133,7 +133,7 @@ defmodule BankWeb.UserAuthTest do
     end
 
     test "does not authenticate if data is missing", %{conn: conn, user: user} do
-      _ = Users.generate_user_session_token(user)
+      _ = UsersSessions.generate_user_session_token(user)
       conn = UserAuth.fetch_current_user(conn, [])
       refute get_session(conn, :user_token)
       refute conn.assigns.current_user
@@ -142,7 +142,7 @@ defmodule BankWeb.UserAuthTest do
 
   describe "on_mount :mount_current_user" do
     test "assigns current_user based on a valid user_token", %{conn: conn, user: user} do
-      user_token = Users.generate_user_session_token(user)
+      user_token = UsersSessions.generate_user_session_token(user)
 
       session =
         conn
@@ -181,7 +181,7 @@ defmodule BankWeb.UserAuthTest do
 
   describe "on_mount :ensure_authenticated" do
     test "authenticates current_user based on a valid user_token", %{conn: conn, user: user} do
-      user_token = Users.generate_user_session_token(user)
+      user_token = UsersSessions.generate_user_session_token(user)
 
       session =
         conn
@@ -226,7 +226,7 @@ defmodule BankWeb.UserAuthTest do
 
   describe "on_mount :redirect_if_user_is_authenticated" do
     test "redirects if there is an authenticated  user ", %{conn: conn, user: user} do
-      user_token = Users.generate_user_session_token(user)
+      user_token = UsersSessions.generate_user_session_token(user)
 
       session =
         conn
